backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-ai-documentintelligence==1.0.0b1
azure-core==1.36.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
dnspython==2.8.0
fastapi==0.109.0
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
isodate==0.7.2
openai==1.12.0
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.6.1
python-dotenv==1.0.0
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.35.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
